Getting separate qualifier dfs for sales uplift and the depletions cross join since sales uplift doesn't require the product_id.
Changing the dims to join on for sales uplift when joining to the venues df and total average.

Adding test cases to cover breaking out by brand, adding a top 10 in filter, and a test case I with data I validated against the DB based off a query they provided.